-Added database commands to schema.sql and data.sql -Fixed Entity classes Added SettingUtil for quick reference. -Made a lot of changes. -Removed items.html and ItemController.java -Added data.sql and schema.sql -Added test.html -Renamed a AccountForm.java and AccountFormValidator.java to AccountRegistrationForm.java and AccountRegistrationValidator.java -Renamed a ItemForm.java and ItemFormValidator.java to ItemSubmissionForm.java and ItemSubmissionFormValidator.java -Renamed ItemInfo.java to OrderInfo.java -Added Authorities.java -Added SettingUtil.java -Updated pom.xml and CSE205GroupProject.iml

+++ server/src/main/java/spring/manager/ItemManager.java
@@ -12,6 +12,7 @@ import spring.repositories.ItemRepository;
 
 import javax.transaction.Transactional;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Optional;
 
 @Service
@@ -34,6 +35,7 @@ public class ItemManager {
     {
         // TODO validate input
 
+
         return itemRepository.save(new Item(name,description,quantity,price));
     }
 
@@ -44,14 +46,14 @@ public class ItemManager {
         return itemRepository.findAll(pageable);
     }
 
-    public Optional<Item> findByItemId( Long id )
+    public Optional<Item> findByItemId( BigInteger id )
     {
         // TODO validate input
 
         return itemRepository.findById(id);
     }
 
-    public Optional<ItemCategories> findItemCategory( Long id )
+    public Optional<ItemCategories> findItemCategory( BigInteger id )
     {
         // TODO validate input
