+++ server/src/main/java/spring/repositories/ItemRepository.java
@@ -4,8 +4,9 @@ package spring.repositories;
 import spring.entity.Item;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
-public interface ItemRepository extends PagingAndSortingRepository<Item,Long> {
-    Optional<Item> findById( Long id );
+public interface ItemRepository extends PagingAndSortingRepository<Item, BigInteger> {
+    Optional<Item> findById( BigInteger id );
 }
